Add files via upload
Added support for Fahrenheit units:

If UNITS=F in config.ini data sent to influxdb will be converted to Fahrenheit, all other values are still in Celcius. If you want to make an adjustment to temphigh or templow you MUST do so in Celcius

Made changes to field names for influxdb:

Temperature now = 1
Humidity now = 2
VPD now = 3

Started pulling temp, humidity, and VPD as an integer instead of a string so that Grafana can properly read data to provide Min, Max and Avg data.

Renaming fields to 1,2,3 was necessary because when creating the fields through the python script the old text values were being sent as a string. InfluxDB will not accept inputs that are not the same type as the original input. Renaming them as 1,2,3 passes them as an int.